Port TinyLFU eviction policy from Caffeine
The implementation included in this change is non-concurrent. It allows
injection of admission policies, but provides none yet. Further, storage
of key/value is left as a separate exercise; this tracks only whether a
key should be cached or evicted. Some small improvements have been made
over Caffeine to reduce allocation of list nodes.