Switch to uint64 keys internally (#25)
Pros: Key length doesn't matter anymore. So helps with memory overhead. Particularly, because we have two maps. One for key-value and one for key-cost.

Cons: Risk of collisions and we can't know when a specific key is evicted (this affects `OnEvict`, but that's something we can work on if it becomes a problem later. Potentially by generating another hash and keeping around to compare).

Changes:
* string to uint64 keys
* onevict uint64