Replace AESHash with MemHash
Currently we are using AESHash(aeshash in runtime) which uses
aesenc instruction. This can crash restritto if it is running
on a hardware not having this instruction. Go runtime, at init
decides to use memhash if this instruction is not available.
memhash is independent of hardware. Also memhash on every call,
checks if aeshash can be calculated(based on the availabilty of
aesenc intruction). This commit replaces AESHash with MemHash.