cache_test.go: TestCacheSetDrops() - remove sync.WaitGroup data race
I /think/ we might have been violating this constraint:
`func (wg *WaitGroup) Add(delta int)...calls with a positive delta
that occur when the counter is zero must happen before a Wait.`
-- https://golang.org/pkg/sync/#WaitGroup.Wait

The goroutine start sychronisation could have been more simply
achieved with just a channel, but I suspected the there might be a
performance reason for originally using a wg so we keep with that
theme.

For what it's worth, the sychronisation mechanism, be it this way,
just a channel, or if it is removed completely, didn't seem to change
my results for this test.

```
WARNING: DATA RACE
Read at 0x00c0003ed878 by goroutine 47:
  sync.(*WaitGroup).Add()
      /home/r/go/src/internal/race/race.go:37 +0x18d
  github.com/dgraph-io/ristretto.TestCacheSetDrops()
      /home/r/ristretto/cache_test.go:206 +0x68
  testing.tRunner()
      /home/r/go/src/testing/testing.go:909 +0x199

Previous write at 0x00c0003ed878 by goroutine 102:
  sync.(*WaitGroup).Wait()
      /home/r/go/src/internal/race/race.go:41 +0xee
  github.com/dgraph-io/ristretto.TestCacheSetDrops.func1()
      /home/r/ristretto/cache_test.go:211 +0x54

Goroutine 47 (running) created at:
  testing.(*T).Run()
      /home/r/go/src/testing/testing.go:960 +0x651
  testing.runTests.func1()
      /home/r/go/src/testing/testing.go:1202 +0xa6
  testing.tRunner()
      /home/r/go/src/testing/testing.go:909 +0x199
  testing.runTests()
      /home/r/go/src/testing/testing.go:1200 +0x521
  testing.(*M).Run()
      /home/r/go/src/testing/testing.go:1117 +0x2ff
  main.main()
      _testmain.go:84 +0x223
```