Chewxy/simd (#227)
Added an unrolled search function for keys. For more comparisons, see github.com/dgraph-io/ristretto-benchmarks, in the searchkeys subdirectory.

The unrolled search function is a hand written, manually optimized assembly function, found in `z/simd/search.s`. The Go equivalent is the function `simd.Clever`, which is retained for clarity. 
Co-authored-by: Manish R Jain <manish@dgraph.io>