Fix(z.Allocator): Make Allocator use Go memory
We are seeing a strange issue when Allocator stores structs in the
memory using C allocated memory. And the same issue shows up even if we
use Go memory, but then re-use it. Without good understanding of why
this happens, the fix for now is to use Go memory and discard it without
reuse.

This PR removes AllocatorPool and switches Allocator to use Go memory.