fix(TTL): ensure expiry buckets are cleaned up
Expiry cleanup is not guaranteed to run at any interval. It attemps to
run roughly twice per bucket duration, but may in fact not run.

If the cleanup does not run at least one per bucket duration, one or
more buckets may be leaked.

This commit adds a field lastCleanedBucketNum to the expirationMap to
keep track of the last bucket that was cleaned up. When cleanup is
called will clean up all the buckets between the last bucket and the
current bucket.