chore(tests): fix flaky test TestDropUpdates (#334)
While submitting a documentation PR, CI failed to run my change with the
following flaky error:

    cache_test.go:866:
          Error Trace:  cache_test.go:866
                              cache_test.go:872
          Error:        Should be true
          Test:         TestDropUpdates

This happens because sometimes sleeping for 1 ms is not enough to ensure
the background processing happens. Replace with c.Wait(), which ensures
it waits until everything is processed.

There is a second race, with the full output below. It is caused when
c.Close() is called with the last Set() still in the buffer. This causes
onEvict() to be called for every item. Call c.Wait() before that point
also.

I was able to reproduce these flakes using my "makeflaky" tool:
https://github.com/evanj/hacks#makeflaky-pause-a-process-to-make-tests-flaky
Running it with the following arguments:

makeflaky -goTest . -race -v -run=TestDropUpdates -count=1000 -failfast