fix(TTL): don't add items into expiration map on update if item has no expiration
If an item was already in the cache, adding it again will cause it to be
added into the expiration map, even if the item has no expiration set.

This effectively is a memory leak. There is no clean up of this "zero
time" bucket, and it will keep growing.