Add Cache.MarshalBinary() and Cache.UnmarshalBinary([]byte) and shardedMap.snapshotEntries() to support Cache.MarshalBinary()
Currently, there is no way to export the cache or load the cache from exported data. This adds the same functionality with the intent of being able to export the cache to an external source such as a replication node or to facilitate cache warming.

Effort was made to ensure concurrency, TTLs, cost, and internal structure were preserved. Test cases for all new functions have been included. README.md was updated with the caveat and considerations that must be taken in order for MarshalBinary() to work properly; this is because MarshalBinary() (and by extension UnmarshalBinary([]byte)) use gob encoding/decoding under the hood. Comments have been added to the relevant new functions. The entire test suite has been executed to ensure sanity.

Also, the spirit (content?) of PR (#354) have been included because they represent a non-zero performance improvement. The one caveat to (#354) is this: it only works if `numShards` is a power of 2. With this in mind, a compile-time check has been added to ensure that `numShards` is a power of 2. Additionally, the compile-time check has comments instructing that the compile-time check must be removed AND `key&shardMask` must be changed back to `key%numShards` should `numShards` need to be changed to a value that is _not_ a power of 2.